+++ accessories/air-purifier.js
@@ -99,7 +99,8 @@ class AirPurifierAccessory extends FanAccessory {
         bind: this,
         props: {
           onData: lockControls,
-          offData: unlockControls
+          offData: unlockControls,
+          setValuePromise: this.performSend.bind(this)
         }
       });
     }
@@ -114,6 +115,7 @@ class AirPurifierAccessory extends FanAccessory {
         props: {
           onData: swingToggle,
           offData: swingToggle,
+          setValuePromise: this.performSend.bind(this)
         }
       });
     }