+++ accessories/fan.js
@@ -11,10 +11,18 @@ class FanAccessory extends SwitchAccessory {
     // Defaults
     config.showSwingMode = config.hideSwingMode === true || config.showSwingMode === false ? false : true;
     config.showRotationDirection = config.hideRotationDirection === true || config.showRotationDirection === false ? false : true;
-    config.stepSize = isNaN(config.stepSize) || config.stepSize > 100 || config.stepSize < 1 ? 1 : config.stepSize
+    config.stepSize = isNaN(config.stepSize) || config.stepSize > 100 || config.stepSize < 1 ? 1 : config.stepSize;
+    
+    if (config.speedSteps) {
+      config.stepSize = Math.floor(100 / config.speedSteps);
+    }
     
     if (config.alwaysResetToDefaults) {
       state.fanSpeed = (config.defaultFanSpeed !== undefined) ? config.defaultFanSpeed : 100;
+
+      if (config.defaultSpeedStep && config.stepSize) {
+        state.fanSpeed = config.defaultSpeedStep * config.stepSize;
+      }
     }
   }
 
@@ -91,6 +99,10 @@ class FanAccessory extends SwitchAccessory {
     if (!this.state.switchState) {
       this.lastFanSpeed = undefined;
     }
+
+    if (config.defaultSpeedStep && config.stepSize) {
+      this.lastFanSpeed = config.defaultSpeedStep * config.stepSize;
+    }
 	  
     // Reset the fan speed back to the default speed when turned off
     if (this.state.switchState === false && config.alwaysResetToDefaults) {
@@ -102,7 +114,7 @@ class FanAccessory extends SwitchAccessory {
   }
 
   async setFanSpeed (hexData) {
-    const { data, host, log, state, name, debug} = this;
+    const { config, data, host, log, state, name, debug} = this;
 
     this.reset();
 
@@ -118,6 +130,12 @@ class FanAccessory extends SwitchAccessory {
       foundSpeeds.push(parts[1])
     })
 
+    if (config.speedCycle && config.speedSteps) {
+      for (let i = 1; i <= config.speedSteps; i++) {
+        foundSpeeds.push(config.stepSize * i);
+      }
+    }
+
     if (foundSpeeds.length === 0) {
 
       return log(`${name} setFanSpeed: No fan speed hex codes provided.`)
@@ -131,10 +149,38 @@ class FanAccessory extends SwitchAccessory {
       return;
     }
 
-    this.lastFanSpeed = closest;
-
     // Get the closest speed's hex data
     hexData = data[`fanSpeed${closest}`];
+
+    if (config.speedCycle) {
+      let fanSpeedHexData = data['fanSpeed'];
+      let fanSpeed = this.lastFanSpeed;
+      hexData = [];
+
+      if (typeof fanSpeedHexData === 'string') {
+        fanSpeedHexData = {
+          data: fanSpeedHexData
+        }
+      }
+
+      if (fanSpeed > closest) {
+        while (fanSpeed < config.speedSteps * config.stepSize) {
+          hexData.push(fanSpeedHexData);
+          fanSpeed += config.stepSize;
+        }
+
+        fanSpeed = 0;
+      }
+
+      if (fanSpeed < closest) {
+        while (fanSpeed < closest) {
+          hexData.push(fanSpeedHexData);
+          fanSpeed += config.stepSize;
+        } 
+      }
+    }
+
+    this.lastFanSpeed = closest;
 
     await this.performSend(hexData);
 
@@ -172,6 +218,7 @@ class FanAccessory extends SwitchAccessory {
         props: {
           onData: swingToggle,
           offData: swingToggle,
+          setValuePromise: this.performSend.bind(this)
         }
       });
     }
@@ -186,7 +233,7 @@ class FanAccessory extends SwitchAccessory {
         setValuePromise: this.setFanSpeed.bind(this),
 		    minStep: config.stepSize,
 		    minValue: 0,
-		    maxVlue: 100
+		    maxValue: 100
       }
     });
 
@@ -199,7 +246,8 @@ class FanAccessory extends SwitchAccessory {
         bind: this,
         props: {
           onData: counterClockwise,
-          offData: clockwise
+          offData: clockwise,
+          setValuePromise: this.performSend.bind(this)
         }
       });
     }